+++ src/routes/ndvi_router.py
